Added exceptions and deleted a doublette in edit-method

+++ src/Controller/ActionController.php
@@ -10,6 +10,7 @@ use App\Form\Type\ActionType;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use function PHPUnit\Framework\throwException;
 
 class ActionController extends AbstractController
 {
@@ -18,11 +19,15 @@ class ActionController extends AbstractController
     public function getAllActions(EntityManagerInterface $entityManager): Response{
         $actions = $entityManager->getRepository(Action::class)->findAll();
 
+        if(!$actions){
+            throw $this->createNotFoundException(("Actions not found"));
+        }
+
         return $this->render('todoList.html.twig',
             ['actions' => $actions]);
     }
 
-    #[Route('/actions/add', name: 'add_action', methods: ['Get','Post'])]
+    #[Route('/actions/add', name: 'add_action', methods: ['GET','POST'])]
     public function addAction(Request $request, EntityManagerInterface $entityManager): Response{
 
         $action = new Action();
@@ -48,6 +53,10 @@ class ActionController extends AbstractController
 
         $action = $entityManager->getRepository(Action::class)->find($id);
 
+        if(!$action){
+            throw $this->createNotFoundException(("Action not found"));
+        }
+
         $form = $this->createForm(ActionType::class, $action);
         $form->handleRequest($request);
 
@@ -64,31 +73,14 @@ class ActionController extends AbstractController
         ]);
     }
 
-    #[Route('/actions/update', name: 'update', methods: ['Get','Post'])]
-    public function updateAction(EntityManagerInterface $entityManager, Action $action,
-                               Request $request): Response{
-
-
-        $form = $this->createForm(ActionType::class, $action);
-        $form->handleRequest($request);
-
-        if($form->isSubmitted() && $form->isValid()){
-            $action = $form->getData();
-            $entityManager->persist($action);
-            $entityManager->flush();
-            return $this->redirectToRoute('get_actions');
-        }
-
-        return $this->render('actions/update.html.twig', [
-            'form' => $form->createView(),
-            'action' => $action
-        ]);
-    }
-
-    #[Route('/actions/remove/{id}', name: 'remove_action', methods: 'Get')]
+    #[Route('/actions/remove/{id}', name: 'remove_action', methods: 'GET')]
     public function removeAction(EntityManagerInterface $entityManager, int $id): Response{
 
         $action = $entityManager->getRepository(Action::class)->find($id);
+
+        if(!$action){
+            throw $this->createNotFoundException(("Action not found"));
+        }
 
         $entityManager->remove($action);
         $entityManager->flush();
